Convert client.py to package
