+++ bika/lims/browser/client/configure.zcml
