+++ bika/lims/browser/client/views/attachments.py
@@ -1,0 +1,78 @@
+from bika.lims import bikaMessageFactory as _
+from bika.lims.browser.bika_listing import BikaListingView
+from plone.app.layout.globals.interfaces import IViewView
+from Products.CMFCore.utils import getToolByName
+from zope.interface import implements
+
+
+class ClientAttachmentsView(BikaListingView):
+    implements(IViewView)
+
+    def __init__(self, context, request):
+        super(ClientAttachmentsView, self).__init__(context, request)
+        self.contentFilter = {'portal_type': 'Attachment',
+                              'sort_order': 'reverse'}
+        self.context_actions = {}
+        self.show_sort_column = False
+        self.show_select_row = False
+        self.show_select_column = True
+        self.pagesize = 50
+        self.form_id = "attachments"
+
+        self.icon = self.portal_url + \
+                    "/++resource++bika.lims.images/attachment_big.png"
+        self.title = self.context.translate(_("Attachments"))
+        self.description = ""
+
+        self.columns = {
+            'getTextTitle': {'title': _('Request ID')},
+            'AttachmentFile': {'title': _('File')},
+            'AttachmentType': {'title': _('Attachment Type')},
+            'ContentType': {'title': _('Content Type')},
+            'FileSize': {'title': _('Size')},
+            'DateLoaded': {'title': _('Date Loaded')},
+        }
+        self.review_states = [
+            {'id': 'default',
+             'title': 'All',
+             'contentFilter': {},
+             'columns': ['getTextTitle',
+                         'AttachmentFile',
+                         'AttachmentType',
+                         'ContentType',
+                         'FileSize',
+                         'DateLoaded']},
+        ]
+
+    def lookupMime(self, name):
+        mimetool = getToolByName(self, 'mimetypes_registry')
+        mimetypes = mimetool.lookup(name)
+        if len(mimetypes):
+            return mimetypes[0].name()
+        else:
+            return name
+
+    def folderitems(self):
+        items = BikaListingView.folderitems(self)
+        for x in range(len(items)):
+            if not items[x].has_key('obj'):
+                continue
+            obj = items[x]['obj']
+            obj_url = obj.absolute_url()
+            file = obj.getAttachmentFile()
+            icon = file.getBestIcon()
+
+            items[x]['AttachmentFile'] = file.filename()
+            items[x]['AttachmentType'] = obj.getAttachmentType().Title()
+            items[x]['AttachmentType'] = obj.getAttachmentType().Title()
+            items[x]['ContentType'] = self.lookupMime(file.getContentType())
+            items[x]['FileSize'] = '%sKb' % (file.get_size() / 1024)
+            items[x]['DateLoaded'] = obj.getDateLoaded()
+
+            items[x]['replace']['getTextTitle'] = \
+                "<a href='%s'>%s</a>" % (obj_url, items[x]['getTextTitle'])
+
+            items[x]['replace']['AttachmentFile'] = \
+                "<a href='%s/at_download/AttachmentFile'>%s</a>" % \
+                (obj_url, items[x]['AttachmentFile'])
+        return items
